refactor(Struct): 컨트롤러 패키지 위치 변경
