+++ src/main/java/com/stempo/api/domain/presentation/controller/AuthController.java
@@ -1,4 +1,4 @@
-package com.stempo.api.domain.presentation;
+package com.stempo.api.domain.presentation.controller;
 
 import com.stempo.api.domain.application.service.AuthService;
 import com.stempo.api.domain.presentation.dto.request.AuthRequestDto;
