+++ src/main/java/com/stempo/api/domain/presentation/controller/HomeworkController.java
@@ -1,4 +1,4 @@
-package com.stempo.api.domain.presentation;
+package com.stempo.api.domain.presentation.controller;
 
 import com.stempo.api.domain.application.service.HomeworkService;
 import com.stempo.api.domain.presentation.dto.request.HomeworkRequestDto;
